Cambios para que el contenido del archivo se conserve aunque se cierre la aplicacion y no se sobrescriba cada vez que empieza la aplicacion, se puede editar el archivo en tiempo real. Cada vez que se le da al boton se guardara el texto

+++ app/src/main/java/com/example/persistencia/MainActivity.java
@@ -26,11 +26,13 @@ public class MainActivity extends AppCompatActivity {
         try{
             File dir= getFilesDir();
             File archivo=new File(dir,"dades.txt");
-            FileWriter escribir=new FileWriter(archivo);
-            String mensaje="Lorem Ipsum is simply dummy text of the printing and typesetting industry.\n";
-            escribir.write(mensaje);
-            escribir.flush();
-            escribir.close();
+            if (!archivo.exists()) {
+                FileWriter escribir = new FileWriter(archivo);
+                String mensaje = "Lorem Ipsum is simply dummy text of the printing and typesetting industry.\n";
+                escribir.write(mensaje);
+                escribir.flush();
+                escribir.close();
+            }
             actualizarEditText();
             b.setOnClickListener(new View.OnClickListener() {
                 public void onClick(View v) {
